requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0